Move more methods to food.cpp
