+++ food.hpp
@@ -3,8 +3,21 @@
 namespace elona
 {
 
+enum class curse_state_t;
+
+void continuous_action_eating();
+void continuous_action_eating_finish();
+
+int chara_anorexia(int = 0);
+void chara_vomit(int = 0);
+void eatstatus(curse_state_t, int);
+void sickifcursed(curse_state_t, int, int);
+void cure_anorexia(int = 0);
+void label_2162();
 void eat_rotten_food();
+
 void cook();
+
 void apply_general_eating_effect();
 void eating_effect_eat_iron();
 void eating_effect_insanity();
@@ -34,6 +47,5 @@ void eating_effect_strength2();
 void eating_effect_will();
 void eating_effect_quick();
 void eating_effect_pregnant();
-void chara_vomit(int = 0);
 
 }
