+++ variables.hpp
@@ -579,7 +579,6 @@ int cell_featread(int = 0, int = 0, int = 0);
 int cell_findspace(int = 0, int = 0, int = 0);
 int cell_itemlist(int = 0, int = 0);
 std::pair<int, int> cell_itemoncell(const position_t& pos);
-int chara_anorexia(int = 0);
 bool chara_unequip(int);
 int characreate(int = 0, int = 0, int = 0, int = 0);
 int clientguide(int = 0);
@@ -885,7 +884,6 @@ void create_pcpic(int, bool);
 void csfix();
 void cslinedown();
 void cslineup();
-void cure_anorexia(int = 0);
 void cursor_check();
 void cutname(std::string&, int = 0);
 void del_chara(int = 0);
@@ -904,7 +902,6 @@ void dmgcard(int = 0, int = 0);
 void dmgplayer(int = 0, int = 0);
 void draw_emo(int = 0, int = 0, int = 0);
 void drawmenu(int = 0);
-void eatstatus(curse_state_t, int);
 void efllistadd(int = 0, int = 0, int = 0, int = 0, int = 0, int = 0);
 void egoadd(int = 0, int = 0);
 void encremove(int = 0, int = 0, int = 0);
@@ -1235,9 +1232,6 @@ void spot_fishing();
 void spot_material();
 void spot_digging();
 void spot_mining_or_wall();
-void continuous_action_eating();
-void continuous_action_eating_finish();
-void label_2162();
 void label_2187();
 void label_2188();
 turn_result_t do_throw_command();
@@ -1412,7 +1406,6 @@ void showcardpile();
 void showeconomy(int, int, const std::string&, int = 0, int = 0);
 void showscroll(const std::string&, int, int, int, int);
 void showtitle(const std::string&, const std::string&, int = 0, int = 0);
-void sickifcursed(curse_state_t, int, int);
 void skillgain(int = 0, int = 0, int = 0, int = 0);
 void skillinit(int = 0, int = 0, int = 0);
 std::string snarmor(const std::string&);
